test/e2e/memcached_test.go: Add metrics test
The test gets the metrics data via the proxy and checks the metrics
are present. It also checks the lint of metrics using the prometheus
linter.

Note: There is an invalid metric in upstream Kubernetes, which is fixed
in 1.14 release. We choose to ignore that one, until we migrate to 1.14
release.